separating out common logic
